Retry if connections fail or ECONNRESET
This change makes the pool more robust by adding a `retries` argument to
`pool_options` (default `2`), which tells the pool to retry a function call if
the connection is closed or encounters an `ECONNRESET` mid-call.

Other minor changes: upgrade to `lodash` instead of `underscore`; add `__id`
field to connections to help with debugging.